// src/app/Interfaces/user.ts
export class User {
    id?: string;
    username?: string;
    role?: string; 
    location?: string; 
    firstName?: string; 
    lastName?: string; 
    profession?: string;
    email?: string; 
    password?: string; 
    country?: string;
    number?: string;
    date: any; 
    imageUrl?: string;


}